+++ qsprpred/about.py
@@ -1,3 +1,3 @@
 """Version of QSPRpred."""
 
-VERSION = "1.3.1"
+VERSION = "1.4.0.dev0"