+++ qsprpred/data/utils/smiles_standardization.py
@@ -6,6 +6,28 @@ from rdkit import Chem
 from rdkit.Chem.SaltRemover import SaltRemover
 from chembl_structure_pipeline import standardizer as chembl_stand
 
+from qsprpred.logs import logger
+
+
+def check_smiles_valid(smiles, throw=True):
+    is_valid = True
+    exception = None
+    if not smiles:
+        is_valid = False
+        exception = ValueError(f"Empty molecule: {smiles}")
+    try:
+        mol = Chem.MolFromSmiles(smiles)
+        if not mol:
+            raise ValueError(f"Invalid molecule: {smiles}")
+        Chem.SanitizeMol(mol)
+    except Exception as exp:
+        is_valid = False
+        exception = exp
+
+    if exception and throw:
+        raise exception
+    else:
+        return is_valid
 
 def neutralize_atoms(mol):
     """Neutralize charged molecules by atom.
@@ -46,14 +68,14 @@ def chembl_smi_standardizer(smi: str, isomericSmiles:bool=True, sanitize:bool=Tr
     """    
     try:
         mol = Chem.MolFromSmiles(smi)
-    except:  # noqa E722
-        print('Could not parse smiles: ', smi)
+        standard_mol = chembl_stand.standardize_mol(mol, sanitize=sanitize)
+        standard_smiles = Chem.MolToSmiles(
+            standard_mol, kekuleSmiles=False, canonical=True, isomericSmiles=isomericSmiles
+        )
+        return standard_smiles
+    except Exception as exp:  # noqa E722
+        logger.exception(f"Could not standardize SMILES: {smi} due to {exp}.")
         return None
-    standard_mol = chembl_stand.standardize_mol(mol, sanitize=sanitize)
-    standard_smiles = Chem.MolToSmiles(
-        standard_mol, kekuleSmiles=False, canonical=True, isomericSmiles=isomericSmiles
-    )
-    return standard_smiles
 
 
 def old_standardize_sanitize(smi: str) -> str: